parameterized gmm test beautify
