+++ test/cpp/modules/manual/ModuleTest.h
@@ -1,0 +1,13 @@
+#include <gtest/gtest.h>
+#include "../../../../src/cpp/runner/ModuleLoader.h"
+
+using ::testing::TestWithParam;
+
+class ModuleTest : public TestWithParam<const char*> {
+public:
+    ~ModuleTest() override;
+    void SetUp() override;
+    void TearDown() override;
+protected:
+    ModuleLoader* moduleLoader;
+};
